pyproject.toml
[tool.poetry]
name = "bsscrapy"
version = "0.1.0"
description = ""
authors = ["jumozaga <jumozaga@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
pandas = "^2.0.0"
sweetviz = "^2.1.4"
pymongo = "^4.3.3"
airflow = "^2.2.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
